overhaul of file hierarchy
**fs.rs**
- creative fs.rs and moved all file system related functions there
**auth.rs**
- removed dead code on lib/auth.rs
**lreywen.rs**
- created library for high level functions that are non essential to RevX2
**main.rs**
- reorganised mod imports
- removed unless dependancies
- removed file system functions
- renamed some files with rev_
- removed all RevX2 functions
