+++ src/lib/auth.rs
@@ -1,5 +1,4 @@
 use serde::{Deserialize, Serialize};
-use validator::Validate;
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct Auth {
