+++ src/main.rs
@@ -1,105 +1,43 @@
-// serde
-//use serde::{Deserialize, Serialize};
-use serde_json::{Result};
+// quark structs
+mod lib {
+    pub mod auth;
+    pub mod message;
+    pub mod user;
+    pub mod br;
+}
 
+use crate::lib::{
+    message::{RMessage, RMessagePayload, Masquerade},
+    br::BrConf,
+    auth::Auth,
+};
 
-// structs
-#[path = "./lib/message.rs"]
-mod message;
-use message::*;
+// reywen fs
+mod fs;
+use fs::{auth_init, bridge_init};
 
-#[path = "./lib/auth.rs"]
-mod auth;
-use auth::*;
-
-
-#[path = "./lib/user.rs"]
-mod user;
-use user::*;
-
-#[path = "./lib/br.rs"]
-mod br;
-use br::*;
-
-
-// non functional - issue #18
 // RevX2
-//mod rev_x;
-//use rev_x::*;
+pub mod rev_x;
+use rev_x::*;
 
-
-// misc
-use rand::Rng;
+// reywen lib
+mod lreywen;
+use lreywen::*;
 
 // network
 use futures_util::{StreamExt, SinkExt};
 use tokio_tungstenite::{connect_async, tungstenite::protocol::Message};
 
-// std
-use std::{
-    io::Read, 
-    fs::File, 
-//    thread, 
-//    time, 
-    str::from_utf8
-};
 
+use std::str::from_utf8;
 use tokio;
-
-
-// import and deserialize auth.conf
-fn conf_file() -> Result<Auth> {
-
-    let mut config_json = File::open("auth.json")
-        .expect("File not found");
-
-    let mut data_str = String::new();
-
-     config_json.read_to_string(&mut data_str)
-        .expect("Error while reading file");
-
-     let conf: Auth = serde_json::from_str(&data_str).expect("failed to interpret conf");
-
-     Ok(conf)
-}
-
-// deserialize messages from ws
-fn message_in(raw: String) -> Result<RMessage> {
-
-
-    let message: Result<RMessage> = serde_json::from_str(&raw);
-
-    match message {
-        Err(rmessage) => Err(rmessage),
-        Ok(ref _rmessage) =>  Ok(message.unwrap())
-    }
-}
-
-
-// import and deserialize bridge.json
-fn bridge_init() -> Result<BrConf> {
-    
-    let mut config_json = File::open("bridge.json")
-        .expect("bridge config file not found");
-
-    let mut brconf_str = String::new();
-    config_json.read_to_string(&mut brconf_str)
-        .expect("Error while reading file");
-
-
-    let config: BrConf = serde_json::from_str(&brconf_str).expect("failed to interpret brconf");
-
-    Ok(config)
-}
-
 
 #[tokio::main]
 async fn main()  {
 
-
     println!("booting...");
     // auth files
-    let data_in = conf_file();
+    let data_in = auth_init();
     let data = match data_in {
         Ok(auth) => auth,
         Err(error) => panic!("Invalid credentials, {error}")
@@ -143,6 +81,7 @@ async fn main()  {
 }
 
 // establishes websocket connection
+
 pub async fn websocket(url: String, authen: Auth, br: BrConf) {
 
      let (ws_stream, _response) = connect_async(url).await.expect("Failed to connect");
@@ -175,20 +114,19 @@ pub async fn websocket(url: String, authen: Auth, br: BrConf) {
 // bridge and message processing
 pub async fn newmain(authen: Auth, out: String, br: BrConf) {
 
-    let inval_message = message_in(out.clone());
-    let inval2 = message_in(out.clone());
-    
+    let inval_message = rev_message_in(out.clone());
+    let inval2 = rev_message_in(out.clone());
+
     match inval_message {
         Err(_) => return,
         Ok(_) => print!("")
 
     };
 
-    // removed error prone characters
-    let message = message_clean(inval_message.expect("failed to process message"));
     
+    let message = inval_message;
     
-    message_process(authen.clone(), message).await;
+    message_process(authen.clone(), message.expect("failed to process message")).await;
     
     // br does not require 'cleaned' messages
     if br.enabled == true {
@@ -260,37 +198,11 @@ pub async fn br_main(auth: Auth, input_message: RMessage, br: BrConf) {
     let payload = RMessagePayload {
         content: message.content.clone(),
         attachments: None,
-        replies: wstoapi_reply(input_message.replies).await,
+        replies: rev_convert_reply(input_message.replies).await,
         masquerade: Some(br_masq),
     };
 
     rev_send(auth, message, payload).await;
-
-}
-
-pub async fn rev_user(auth: Auth, target: String)   -> Result<RUserFetch> {
-
-  //  println!("rev: user");
-   
-    let client: std::result::Result<reqwest::Response, reqwest::Error> =
-    reqwest::Client::new()
-    .get(format!("https://api.revolt.chat/users/{target}"))
-    .header("x-bot-token", auth.token)
-    .send().await;
-
-    let client_res = match client {
-        Ok(_) => client.unwrap().text().await.unwrap(),
-        Err(_) => "Err:\n{error}".to_string()
-    };
-
-
-  //  println!("{:?}", client_res);
-
-    let message: Result<RUserFetch> = serde_json::from_str(&client_res);
-    match message {
-        Ok(_) => return Ok(message.unwrap()),
-        Err(_) => return message
-    };
 
 }
 
@@ -306,7 +218,7 @@ pub async fn message_process(data: Auth, message_in: RMessage) {
         return
     };
 
-    let message = message_clean(message_in);
+    let message = rev_message_clean(message_in).await;
 
     let content_vec =  content.as_ref().expect("failed to split vec").split(' ').collect::<Vec<&str>>();
 
@@ -328,6 +240,7 @@ pub async fn message_process(data: Auth, message_in: RMessage) {
 }
 
 // masq wrapper for rev_send
+
 pub async fn sendas(auth: Auth, message: RMessage, content_vec: Vec<&str>) {
 
     if content_vec.len() < 3 {
@@ -357,7 +270,7 @@ pub async fn sendas(auth: Auth, message: RMessage, content_vec: Vec<&str>) {
     };
 
  
-    let replier = wstoapi_reply(message.replies.clone()).await;
+    let replier = rev_convert_reply(message.replies.clone()).await;
 
     let returner = RMessagePayload {
           content: Some(content),
@@ -368,124 +281,4 @@ pub async fn sendas(auth: Auth, message: RMessage, content_vec: Vec<&str>) {
 
     rev_send(auth.clone(), message.clone(), returner).await;
     rev_del(auth.clone(), message.clone()).await;
-}
-// converts websocket replies to API compatible replies
-pub async fn wstoapi_reply(input: Option<Vec<String>>) -> Option<Vec<RReplies>> {
-
-    if input == None {
-        
-        return None
-    
-    }else {
-        
-        let mut repstruct = vec![];
-        let iter = input.clone()?.len();
-
-        for x in 0..iter {
-            
-            let input_iter = &input.as_ref().expect("failed to convert input wstoapi")[x];
-            
-            let reply = RReplies {
-                id: input_iter.to_string(),
-                mention: false,
-            };
-            repstruct.push(reply);
-        };
-
-        return Some(repstruct)
-    };
-
-}
-
-// non masq wrapper for rev_send
-pub async fn send(auth: Auth, message: RMessage, content: String) {
-
-    let reply = RReplies {
-        id: message._id.clone(),
-        mention: false,
-    };
-    let payload2 = RMessagePayload {
-        content: Some(content),
-        replies: Some(vec![reply]),
-          attachments: None,
-          masquerade: None
-    };
-
-    rev_send(auth, message, payload2).await;
-
-}
-
-
-// deletes messages over http
-pub async fn rev_del(auth: Auth, message: RMessage) {
-
-    let channel = message.channel;
-    let target = message._id;
-
-    let client: std::result::Result<reqwest::Response, reqwest::Error> =
-    reqwest::Client::new()
-    .delete(format!("https://api.revolt.chat/channels/{channel}/messages/{target}"))
-    .header("x-bot-token", auth.token)
-    .send().await;
-
-     match client {
-        Ok(_) => return,
-        Err(_) => println!("Err:\n{:?}", client)
-    };
-
-
-}
-// sends messages over http
-pub async fn rev_send(auth: Auth, message: RMessage, payload: RMessagePayload)  {
-    
-    let channel = message.channel;
-
-    let mut random = rand::thread_rng();
-    let idem: i64 = random.gen();
-
-    let payload2 = serde_json::to_string(&payload).unwrap();
-
-    let client: std::result::Result<reqwest::Response, reqwest::Error> = 
-        reqwest::Client::new()
-        .post(format!("https://api.revolt.chat/channels/{channel}/messages"))
-        .header("x-bot-token", auth.token)
-        .header("Idempotency-Key", idem)
-        .header("Content-Type", "application/json")
-        .body(payload2)
-        .send().await;
- 
-    match client {
-        //Ok(_) => println!("{}", client.unwrap().text().await.unwrap()),
-        Ok(_) => return,
-        Err(_) => println!("Err:\n{:?}", client)
-    };
-}
-
-
-// cleans invalid characters such as \n and \
-pub fn message_clean(mut message: RMessage) -> RMessage {
-
-    if message.content == None {
-        return message
-    };
-
-    let mut out = String::new();
-
-    let iter = message.content.as_ref().unwrap().chars().count();
-    let content = message.content.as_ref().unwrap();
-
-
-    for x in 0..iter {
-        let current = content.chars().nth(x);
-
-        if current == Some('\n') {
-            out += "\\n";
-        }else if current == Some('\\') {
-            out += "\\\\";
-        }else {
-            out += &current.unwrap().to_string();  
-        };
-    };
-    message.content = Some(out);
-    return message    
 }
